requirements.txt
Flask
requests
yfinance
flask_cors
pyyaml
supervisor